+++ Apps/Examples/Examples/All Examples/AdvancedViewportGesturesExample.swift
@@ -36,6 +36,7 @@ final class AdvancedViewportGesturesExample: UIViewController, ExampleProtocol {
 
         mapView.location.options.puckType = .puck2D(.makeDefault(showBearing: true))
         mapView.location.options.puckBearing = .course
+        mapView.location.options.puckBearingEnabled = true
 
         followPuckViewportState = mapView.viewport.makeFollowPuckViewportState(
             options: FollowPuckViewportStateOptions(